refactor: ChatBubbleMakable 프로토콜 구현 및 분기 처리
